pyproject.toml
[tool.poetry]
name = "precisely"
version = "0.1.0"
description = "API Client for recruitement test"
authors = ["Gustaf Hallberg <ghallberg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
